refact: move provide to visitors
