+++ crates/mako/src/transform.rs
@@ -20,7 +20,6 @@ use crate::ast_2::file::File;
 use crate::compiler::Context;
 use crate::module::ModuleAst;
 use crate::targets;
-use crate::transformers::transform_provide::Provide;
 use crate::transformers::transform_px2rem::Px2Rem;
 use crate::transformers::transform_react::mako_react;
 use crate::visitors::css_assets::CSSAssets;
@@ -28,6 +27,7 @@ use crate::visitors::css_flexbugs::CSSFlexbugs;
 use crate::visitors::default_export_namer::DefaultExportNamer;
 use crate::visitors::dynamic_import_to_require::DynamicImportToRequire;
 use crate::visitors::env_replacer::{build_env_map, EnvReplacer};
+use crate::visitors::provide::Provide;
 use crate::visitors::try_resolve::TryResolve;
 use crate::visitors::virtual_css_modules::VirtualCSSModules;
