+++ crates/mako/src/transformers/mod.rs
@@ -5,7 +5,6 @@ pub mod transform_mako_require;
 pub mod transform_meta_url_replacer;
 pub mod transform_optimize_define_utils;
 // pub mod transform_optimize_package_imports;
-pub mod transform_provide;
 pub mod transform_px2rem;
 pub mod transform_react;
 pub mod utils;
