+++ crates/mako/src/visitors/mod.rs
@@ -6,5 +6,6 @@ pub(crate) mod default_export_namer;
 pub(crate) mod dep_analyzer;
 pub(crate) mod dynamic_import_to_require;
 pub(crate) mod env_replacer;
+pub(crate) mod provide;
 pub(crate) mod try_resolve;
 pub(crate) mod virtual_css_modules;
